Problems handling the diskStorage with multer still pending

+++ Controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require("../models/product");
+const fileHelper = require('../utils/file');
 const { validationResult } = require('express-validator');
 
 
@@ -9,7 +10,7 @@ exports.getProducts = (req, res, next) => {
         prods: products,
         pageTitle: "All product",
         path: "/admin/products"
-      });
+      })
     })
     .catch((err) => {
       console.log(err);
@@ -29,13 +30,11 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    console.log(errors.array());
+  if (!image) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
       path: '/admin/edit-product',
@@ -43,11 +42,28 @@ exports.postAddProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: title,
-        imageUrl: imageUrl,
         price: price,
         description: description
       },
-      errorMessage: errors.array()[0].msg,
+      errorMessage: 'the image is not set',
+      validationErrors: []
+    })
+  }
+  const imageUrl = image.path;
+
+  if (!errors.isEmpty()) {
+    // console.log(errors.array());
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        price: price,
+        description: description
+      },
+      errorMessage: 'errors.array()[0].msg',
       validationErrors: errors.array()
     });
   }
@@ -95,7 +111,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDescription = req.body.description;
   const errors = validationResult(req);
 
@@ -107,7 +123,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError: true,
       product: {
         title: updatedTitle,
-        imageUrl: updatedImageUrl,
         price: updatedPrice,
         description: updatedDescription,
         _id: prodId
@@ -124,7 +139,10 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDescription;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path;
+      }
       product.save()
         .then((result) => {
           // console.log("updated Product");
@@ -138,10 +156,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({ _id: productId, userId: req.user._id })
-    .then(result => {
+  Product.findById(productId)
+    .then(prod => {
+      if (!prod) {
+        return next(new Error('No product found.'))
+      }
+      fileHelper.deleteFile(product.imageUrl)
 
-      res.redirect('/admin/products');
+      return Product.deleteOne({ _id: productId, userId: req.user._id })
+        .then(result => {
+          res.redirect('/admin/products');
+        })
+        .catch(err => {
+          const error = new Error(err);
+          error.httpStatusCode = 500;
+          return next(error);
+        });
     })
-    .catch(err => console.log(err));
 };
